requirements.txt
Flask==0.12.2
numpy==1.13.3
Pillow==4.3.0
tensorflow==1.4.0
